\name{fCalendarMasks}

\alias{MskedObjects}

\alias{datax}
            

\title{Masked Objects}


\description{
    
    A collection and description of functions which
    mask function from R's base installation.
    \cr
    
}


\usage{
datax(\dots, list = character(0), package = NULL, lib.loc = NULL,
    verbose = getOption("verbose"), envir = .GlobalEnv)
}


\arguments{

    \item{list, package, lib.loc, verbose, envir}{
        [data] - \cr
        additional arguments to the \code{data} function.
        }
    \item{\dots}{
        a sequence of names or literal character strings.
        }
        
}


\details{

    For details we refer to the original help pages.
    
}
    


\author{

    Diethelm Wuertz for the Rmetrics \R-port.
}


\keyword{programming}

