\name{CalendarData}

\alias{CalendarData}

\alias{timezones.db}


\title{fCalendar Data Sets}


\description{

    A collection and description of data sets 
    used in the the 'fCalendar' package. This 
    data collection holds the timezone data 
    file with all available 'Ical' time zone 
    names.
    \cr
                
    The timezone datafile is:
    
    \tabular{ll}{
    \code{timezones.db} \tab file with all available Ical time zone names. }
    
}


\details{
    
    \bold{Calendar Data:}
    \cr\cr
    \code{holidays.db.R} is the database of world wide holidays with
    informations collected from several sources on the internet.
    \cr
    \code{timezones.csv} is a list of timezones information extracted
    from Olsen's database. 
    
}


\source{

    The database was obtained from Downloadable from
    \cr
    \emph{ftp://elsie.nci.nih.gov/pub/}

    
}


\examples{
## SOURCE("fCalendar.25A-CalendarData")

}


\keyword{datasets}

