\name{ISO8601Standard}

\alias{ISO8601Standard}


\title{Summary of ISO8601 Standard}


\description{
        
    A summary and description how Rmetrics uses the
    ISO8601 standard.
                        
}


\author{

    Diethelm Wuertz for this summary.
    
}


\keyword{programming}

