\name{currentYear}


\title{Current Year}


\alias{currentYear}
\alias{currentMonth}


\description{

    Returns the current year and/or month.

}


\usage{
currentYear
currentMonth
}


\value{

    returns the current year or month, a numeric variable.

}


\note{

    It is not allowed to change this variable.
    
}


\examples{
## currentYear -
   currentYear
   
## currentMonth -
   currentMonth
}


\keyword{chron}

