\name{DaylightSavingTime}


\alias{DaylightSavingTime}


\alias{Abidjan}
\alias{Accra}
\alias{Adak}
\alias{Addis_Ababa}
\alias{Adelaide}
\alias{Aden}
\alias{Algiers}
\alias{Almaty}
\alias{Amman}
\alias{Amsterdam}
\alias{Anadyr}
\alias{Anchorage}
\alias{Andorra}
\alias{Anguilla}
\alias{Antananarivo}
\alias{Antigua}
\alias{Apia}
\alias{Aqtau}
\alias{Aqtobe}
\alias{Araguaina}
\alias{Aruba}
\alias{Ashgabat}
\alias{Asmara}
\alias{Asuncion}
\alias{Athens}
\alias{Atikokan}
\alias{Auckland}
\alias{Azores}
\alias{Baghdad}
\alias{Bahia}
\alias{Bahrain}
\alias{Baku}
\alias{Bamako}
\alias{Bangkok}
\alias{Bangui}
\alias{Banjul}
\alias{Barbados}
\alias{Beirut}
\alias{Belem}
\alias{Belgrade}
\alias{Belize}
\alias{Berlin}
\alias{Bermuda}
\alias{Bishkek}
\alias{Bissau}
\alias{Blanc-Sablon}
\alias{Blantyre}
\alias{Boa_Vista}
\alias{Bogota}
\alias{Boise}
\alias{Bratislava}
\alias{Brazzaville}
\alias{Brisbane}
\alias{Broken_Hill}
\alias{Brunei}
\alias{Brussels}
\alias{Bucharest}
\alias{Budapest}
\alias{Buenos_Aires}
\alias{BuenosAires}
\alias{Bujumbura}
\alias{Cairo}
\alias{Calcutta}
\alias{Cambridge_Bay}
\alias{Campo_Grande}
\alias{Canary}
\alias{Cancun}
\alias{Cape_Verde}
\alias{Caracas}
\alias{Casablanca}
\alias{Casey}
\alias{Catamarca}
\alias{Cayenne}
\alias{Cayman}
\alias{Center}
\alias{Ceuta}
\alias{Chagos}
\alias{Chatham}
\alias{Chicago}
\alias{Chihuahua}
\alias{Chisinau}
\alias{Choibalsan}
\alias{Chongqing}
\alias{Christmas}
\alias{Cocos}
\alias{Colombo}
\alias{Comoro}
\alias{Conakry}
\alias{Copenhagen}
\alias{Cordoba}
\alias{Costa_Rica}
\alias{Cuiaba}
\alias{Curacao}
\alias{Currie}
\alias{Dakar}
\alias{Damascus}
\alias{Danmarkshavn}
\alias{Dar_es_Salaam}
\alias{Darwin}
\alias{Davis}
\alias{Dawson}
\alias{Dawson_Creek}
\alias{Denver}
\alias{Detroit}
\alias{Dhaka}
\alias{Dili}
\alias{Djibouti}
\alias{Dominica}
\alias{Douala}
\alias{Dubai}
\alias{Dublin}
\alias{DumontDUrville}
\alias{Dushanbe}
\alias{Eastern}
\alias{Edmonton}
\alias{Efate}
\alias{Eirunepe}
\alias{El_Aaiun}
\alias{El_Salvador}
\alias{Enderbury}
\alias{Eucla}
\alias{Fakaofo}
\alias{Faroe}
\alias{Fiji}
\alias{Fortaleza}
\alias{Frankfurt}
\alias{Freetown}
\alias{Funafuti}
\alias{Gaborone}
\alias{Galapagos}
\alias{Gambier}
\alias{Gaza}
\alias{Gibraltar}
\alias{Glace_Bay}
\alias{Godthab}
\alias{Goose_Bay}
\alias{Grand_Turk}
\alias{Grenada}
\alias{Guadalcanal}
\alias{Guadeloupe}
\alias{Guam}
\alias{Guatemala}
\alias{Guayaquil}
\alias{Guernsey}
\alias{Guyana}
\alias{Halifax}
\alias{Harare}
\alias{Harbin}
\alias{Havana}
\alias{Helsinki}
\alias{Hermosillo}
\alias{Hobart}
\alias{Hong_Kong}
\alias{HongKong}
\alias{Honolulu}
\alias{Hovd}
\alias{Indianapolis}
\alias{Inuvik}
\alias{Iqaluit}
\alias{Irkutsk}
\alias{Isle_of_Man}
\alias{Istanbul}
\alias{Jakarta}
\alias{Jamaica}
\alias{Jayapura}
\alias{Jersey}
\alias{Jerusalem}
\alias{Johannesburg}
\alias{Johnston}
\alias{Jujuy}
\alias{Juneau}
\alias{Kabul}
\alias{Kaliningrad}
\alias{Kamchatka}
\alias{Kampala}
\alias{Karachi}
\alias{Kashgar}
\alias{Katmandu}
\alias{Kerguelen}
\alias{Khartoum}
\alias{Kiev}
\alias{Kigali}
\alias{Kinshasa}
\alias{Kiritimati}
\alias{Knox}
\alias{Kosrae}
\alias{Krasnoyarsk}
\alias{Kuala_Lumpur}
\alias{KualaLumpur}
\alias{Kuching}
\alias{Kuwait}
\alias{Kwajalein}
\alias{La_Paz}
\alias{La_Rioja}
\alias{Lagos}
\alias{Libreville}
\alias{Lima}
\alias{Lindeman}
\alias{Lisbon}
\alias{Ljubljana}
\alias{Lome}
\alias{London}
\alias{Longyearbyen}
\alias{Lord_Howe}
\alias{Los_Angeles}
\alias{LosAngeles}
\alias{Louisville}
\alias{Luanda}
\alias{Lubumbashi}
\alias{Lusaka}
\alias{Luxembourg}
\alias{Macau}
\alias{Maceio}
\alias{Madeira}
\alias{Madrid}
\alias{Magadan}
\alias{Mahe}
\alias{Majuro}
\alias{Makassar}
\alias{Malabo}
\alias{Maldives}
\alias{Malta}
\alias{Managua}
\alias{Manaus}
\alias{Manila}
\alias{Maputo}
\alias{Marengo}
\alias{Mariehamn}
\alias{Marigot}
\alias{Marquesas}
\alias{Martinique}
\alias{Maseru}
\alias{Mauritius}
\alias{Mawson}
\alias{Mayotte}
\alias{Mazatlan}
\alias{Mbabane}
\alias{McMurdo}
\alias{Melbourne}
\alias{Mendoza}
\alias{Menominee}
\alias{Merida}
\alias{Mexico_City}
\alias{MexicoCity}
\alias{Midway}
\alias{Minsk}
\alias{Miquelon}
\alias{Mogadishu}
\alias{Monaco}
\alias{Moncton}
\alias{Monrovia}
\alias{Monterrey}
\alias{Montevideo}
\alias{Monticello}
\alias{Montreal}
\alias{Montserrat}
\alias{Moscow}
\alias{Muscat}
\alias{Nairobi}
\alias{Nassau}
\alias{Nauru}
\alias{Ndjamena}
\alias{New_Salem}
\alias{New_York}
\alias{NewYork}
\alias{Niamey}
\alias{Nicosia}
\alias{Nipigon}
\alias{Niue}
\alias{Nome}
\alias{Norfolk}
\alias{Noronha}
\alias{Nouakchott}
\alias{Noumea}
\alias{Novosibirsk}
\alias{Omsk}
\alias{Oral}
\alias{Oslo}
\alias{Ouagadougou}
\alias{Pacific}
\alias{Pago_Pago}
\alias{Palau}
\alias{Palmer}
\alias{Panama}
\alias{Pangnirtung}
\alias{Paramaribo}
\alias{Paris}
\alias{Perth}
\alias{Petersburg}
\alias{Phnom_Penh}
\alias{Phoenix}
\alias{Pitcairn}
\alias{Podgorica}
\alias{Ponape}
\alias{Pontianak}
\alias{Port_Moresby}
\alias{Port_of_Spain}
\alias{Port-au-Prince}
\alias{Porto_Velho}
\alias{Porto-Novo}
\alias{Prague}
\alias{Puerto_Rico}
\alias{Pyongyang}
\alias{Qatar}
\alias{Qyzylorda}
\alias{Rainy_River}
\alias{Rangoon}
\alias{Rankin_Inlet}
\alias{Rarotonga}
\alias{Recife}
\alias{Regina}
\alias{Resolute}
\alias{Reunion}
\alias{Reykjavik}
\alias{Riga}
\alias{Rio_Branco}
\alias{Rio_Gallegos}
\alias{Riyadh}
\alias{Rome}
\alias{Rothera}
\alias{Saigon}
\alias{Saipan}
\alias{Sakhalin}
\alias{Samara}
\alias{Samarkand}
\alias{San_Juan}
\alias{San_Marino}
\alias{Santiago}
\alias{Santo_Domingo}
\alias{Sao_Paulo}
\alias{Sao_Tome}
\alias{Sarajevo}
\alias{Scoresbysund}
\alias{Seoul}
\alias{Shanghai}
\alias{Shiprock}
\alias{Simferopol}
\alias{Singapore}
\alias{Skopje}
\alias{Sofia}
\alias{South_Georgia}
\alias{South_Pole}
\alias{St_Barthelemy}
\alias{St_Helena}
\alias{St_Johns}
\alias{St_Kitts}
\alias{St_Lucia}
\alias{St_Thomas}
\alias{St_Vincent}
\alias{Stanley}
\alias{Stockholm}
\alias{Swift_Current}
\alias{Sydney}
\alias{Syowa}
\alias{Tahiti}
\alias{Taipei}
\alias{Tallinn}
\alias{Tarawa}
\alias{Tashkent}
\alias{Tbilisi}
\alias{Tegucigalpa}
\alias{Tehran}
\alias{Tell_City}
\alias{Thimphu}
\alias{Thule}
\alias{Thunder_Bay}
\alias{Tijuana}
\alias{Tirane}
\alias{Tokyo}
\alias{Tongatapu}
\alias{Toronto}
\alias{Tortola}
\alias{Tripoli}
\alias{Truk}
\alias{Tucuman}
\alias{Tunis}
\alias{Ulaanbaatar}
\alias{Urumqi}
\alias{Ushuaia}
\alias{Uzhgorod}
\alias{Vaduz}
\alias{Vancouver}
\alias{Vatican}
\alias{Vevay}
\alias{Vienna}
\alias{Vientiane}
\alias{Vilnius}
\alias{Vincennes}
\alias{Vladivostok}
\alias{Volgograd}
\alias{Vostok}
\alias{Wake}
\alias{Wallis}
\alias{Warsaw}
\alias{Whitehorse}
\alias{Winamac}
\alias{Windhoek}
\alias{Winnipeg}
\alias{Yakutat}
\alias{Yakutsk}
\alias{Yekaterinburg}
\alias{Yellowknife}
\alias{Yerevan}
\alias{Zagreb}
\alias{Zaporozhye}
\alias{Zurich}


\title{Daylight Saving Time Rules}


\description{

    Functions for about 400 cities and regions, which return time 
    zone offsets and daylight saving time rules. 
    
}


\details{

    As a selection of these functions:

    Adelaide Algiers Amsterdam Anchorage Andorra Athens Auckland Bahrain
    Bangkok Beirut Belfast Belgrade Berlin Bogota Bratislava Brisbane
    Brussels Bucharest Budapest BuenosAires Cairo Calcutta Caracas
    Casablanca Cayman Chicago Copenhagen Darwin Denver Detroit Dubai
    Dublin Eastern Edmonton Frankfurt Helsinki HongKong Honolulu
    Indianapolis Istanbul Jakarta Jerusalem Johannesburg Kiev
    KualaLumpur Kuwait Lagos Lisbon Ljubljana London LosAngeles
    Luxembourg Madrid Manila Melbourne MexicoCity Monaco Montreal Moscow
    Nairobi Nassau NewYork Nicosia Oslo Pacific Paris Perth Prague Riga
    Riyadh Rome Seoul Shanghai Singapore Sofia Stockholm Sydney Taipei
    Tallinn Tehran Tokyo Tunis Vaduz Vancouver Vienna Vilnius Warsaw
    Winnipeg Zagreb Zurich, ...
    
}


\note{

    There are currently two synonymes available "Pacific" for Los Angeles
    and "Eastern" for New York. 
    
    Note that we leave the space in all double named cities like New York 
    or Hong Kong and use an underscore for it.
    
    All the entries are retrieved from the tzdata library which is 
    available under GNU's GPL licence. 

}


\examples{
## DST Rules for Zurich:
   head(Zurich())
   tail(Zurich())
}


\keyword{data}

