\name{holidayDates}


\title{Public and Ecclesiastical Holidays}


\alias{holidayDates}

\alias{Septuagesima}
\alias{Quinquagesima}
\alias{AshWednesday}
\alias{PalmSunday}
\alias{GoodFriday}
\alias{EasterSunday}
\alias{EasterMonday}
\alias{RogationSunday}
\alias{Ascension}
\alias{Pentecost}
\alias{PentecostMonday}
\alias{TrinitySunday}
\alias{CorpusChristi}
\alias{ChristTheKing}
\alias{Advent1st}
\alias{Advent2nd}
\alias{Advent3rd}
\alias{Advent4th}
\alias{ChristmasEve}
\alias{ChristmasDay}
\alias{BoxingDay}
\alias{NewYearsDay}
\alias{SolemnityOfMary}
\alias{Epiphany}
\alias{PresentationOfLord}
\alias{Annunciation}
\alias{TransfigurationOfLord}
\alias{AssumptionOfMary}
\alias{BirthOfVirginMary}
\alias{CelebrationOfHolyCross}
\alias{MassOfArchangels}
\alias{AllSaints}
\alias{AllSouls}
\alias{LaborDay}
\alias{CHBerchtoldsDay}
\alias{CHSechselaeuten}
\alias{CHAscension}
\alias{CHConfederationDay}
\alias{CHKnabenschiessen}
\alias{GBMayDay}
\alias{GBBankHoliday}
\alias{GBSummerBankHoliday}
\alias{GBMilleniumDay}
\alias{DEAscension}
\alias{DECorpusChristi}
\alias{DEGermanUnity}
\alias{DEChristmasEve}
\alias{DENewYearsEve}
\alias{FRFetDeLaVictoire1945}
\alias{FRAscension}
\alias{FRBastilleDay}
\alias{FRAssumptionVirginMary}
\alias{FRAllSaints}
\alias{FRArmisticeDay}
\alias{ITEpiphany}
\alias{ITLiberationDay}
\alias{ITAssumptionOfVirginMary}
\alias{ITAllSaints}
\alias{ITStAmrose}
\alias{ITImmaculateConception}
\alias{USDecorationMemorialDay}
\alias{USPresidentsDay}
\alias{USNewYearsDay}
\alias{USInaugurationDay}
\alias{USMLKingsBirthday}
\alias{USLincolnsBirthday}
\alias{USWashingtonsBirthday}
\alias{USMemorialDay}
\alias{USIndependenceDay}
\alias{USLaborDay}
\alias{USColumbusDay}
\alias{USElectionDay}
\alias{USVeteransDay}
\alias{USThanksgivingDay}
\alias{USChristmasDay}
\alias{USCPulaskisBirthday}
\alias{USGoodFriday}
\alias{CAVictoriaDay}
\alias{CAFamilyDay}
\alias{CACanadaDay}
\alias{CACivicProvincialHoliday}
\alias{CALabourDay}
\alias{CAThanksgivingDay}
\alias{CaRemembranceDay}
\alias{JPVernalEquinox}
\alias{JPNewYearsDay}
\alias{JPGantan}
\alias{JPBankHolidayJan2}
\alias{JPBankHolidayJan3}
\alias{JPComingOfAgeDay}
\alias{JPSeijinNoHi}
\alias{JPNatFoundationDay}
\alias{JPKenkokuKinenNoHi}
\alias{JPGreeneryDay}
\alias{JPMidoriNoHi}
\alias{JPConstitutionDay}
\alias{JPKenpouKinenBi}
\alias{JPNationHoliday}
\alias{JPKokuminNoKyujitu}
\alias{JPChildrensDay}
\alias{JPKodomoNoHi}
\alias{JPMarineDay}
\alias{JPUmiNoHi}
\alias{JPRespectForTheAgedDay}
\alias{JPKeirouNOhi}
\alias{JPAutumnalEquinox}
\alias{JPShuubunNoHi}
\alias{JPHealthandSportsDay}
\alias{JPTaiikuNoHi}
\alias{JPNationalCultureDay}
\alias{JPBunkaNoHi}
\alias{JPThanksgivingDay}
\alias{JPKinrouKanshaNoHi}
\alias{JPEmperorsBirthday}
\alias{JPTennouTanjyouBi}
\alias{JPBankHolidayDec31}


\description{
    
    A collection and description of functions and methods 
    dealing with holiday dates in the G7 countries and
    Switzerland. 
    \cr
   
}


\usage{
Septuagesima(year = currentYear)
Quinquagesima(year = currentYear)
AshWednesday(year = currentYear)
PalmSunday(year = currentYear)
GoodFriday(year = currentYear)
EasterSunday(year = currentYear)
EasterMonday(year = currentYear)
RogationSunday(year = currentYear)
Ascension(year = currentYear)
Pentecost(year = currentYear)
PentecostMonday(year = currentYear)
TrinitySunday(year = currentYear)
CorpusChristi(year = currentYear)
ChristTheKing(year = currentYear)
Advent1st(year = currentYear)
Advent2nd(year = currentYear)
Advent3rd(year = currentYear)
Advent4th(year = currentYear)
ChristmasEve(year = currentYear)
ChristmasDay(year = currentYear)
BoxingDay(year = currentYear)
NewYearsDay(year = currentYear)
SolemnityOfMary(year = currentYear)
Epiphany(year = currentYear)
PresentationOfLord(year = currentYear)
Annunciation(year = currentYear)
TransfigurationOfLord(year = currentYear)
AssumptionOfMary(year = currentYear)
BirthOfVirginMary(year = currentYear)
CelebrationOfHolyCross(year = currentYear)
MassOfArchangels(year = currentYear)
AllSaints(year = currentYear)
AllSouls(year = currentYear)
LaborDay(year = currentYear)
CHBerchtoldsDay(year = currentYear)
CHSechselaeuten(year = currentYear)
CHAscension(year = currentYear)
CHConfederationDay(year = currentYear)
CHKnabenschiessen(year = currentYear)
GBMayDay(year = currentYear)
GBBankHoliday(year = currentYear)
GBSummerBankHoliday(year = currentYear)
GBMilleniumDay(year = currentYear)
DEAscension(year = currentYear)
DECorpusChristi(year = currentYear)
DEGermanUnity(year = currentYear)
DEChristmasEve(year = currentYear)
DENewYearsEve(year = currentYear)
FRFetDeLaVictoire1945(year = currentYear)
FRAscension(year = currentYear)
FRBastilleDay(year = currentYear)
FRAssumptionVirginMary(year = currentYear)
FRAllSaints(year = currentYear)
FRArmisticeDay(year = currentYear)
ITEpiphany(year = currentYear)
ITLiberationDay(year = currentYear)
ITAssumptionOfVirginMary(year = currentYear)
ITAllSaints(year = currentYear)
ITStAmrose(year = currentYear)
ITImmaculateConception(year = currentYear)
USDecorationMemorialDay(year = currentYear)
USPresidentsDay(year = currentYear)
USNewYearsDay(year = currentYear)
USInaugurationDay(year = currentYear)
USMLKingsBirthday(year = currentYear)
USLincolnsBirthday(year = currentYear)
USWashingtonsBirthday(year = currentYear)
USMemorialDay(year = currentYear)
USIndependenceDay(year = currentYear)
USLaborDay(year = currentYear)
USColumbusDay(year = currentYear)
USElectionDay(year = currentYear)
USVeteransDay(year = currentYear)
USThanksgivingDay(year = currentYear)
USChristmasDay(year = currentYear)
USCPulaskisBirthday(year = currentYear)
USGoodFriday(year = currentYear)
CAVictoriaDay(year = currentYear)
CACanadaDay(year = currentYear)
CACivicProvincialHoliday(year = currentYear)
CALabourDay(year = currentYear)
CAThanksgivingDay(year = currentYear)
CaRemembranceDay(year = currentYear)
JPVernalEquinox (year = currentYear)
JPNewYearsDay(year = currentYear)
JPGantan(year = currentYear)
JPBankHolidayJan2(year = currentYear)
JPBankHolidayJan3(year = currentYear)
JPComingOfAgeDay(year = currentYear)
JPSeijinNoHi(year = currentYear)
JPNatFoundationDay(year = currentYear)
JPKenkokuKinenNoHi(year = currentYear)
JPGreeneryDay(year = currentYear)
JPMidoriNoHi(year = currentYear)
JPConstitutionDay(year = currentYear)
JPKenpouKinenBi(year = currentYear)
JPNationHoliday(year = currentYear)
JPKokuminNoKyujitu(year = currentYear)
JPChildrensDay(year = currentYear)
JPKodomoNoHi(year = currentYear)
JPMarineDay(year = currentYear)
JPUmiNoHi(year = currentYear)
JPRespectForTheAgedDay(year = currentYear)
JPKeirouNOhi(year = currentYear)
JPAutumnalEquinox(year = currentYear)
JPShuubunNoHi(year = currentYear)
JPHealthandSportsDay(year = currentYear)
JPTaiikuNoHi(year = currentYear)
JPNationalCultureDay(year = currentYear)
JPBunkaNoHi(year = currentYear)
JPThanksgivingDay(year = currentYear)
JPKinrouKanshaNoHi(year = currentYear)
JPEmperorsBirthday(year = currentYear)
JPTennouTanjyouBi(year = currentYear)
JPBankHolidayDec31(year = currentYear)
}


\arguments{

    \item{year}{ 
        an integer value or vector of year numbers including the
        century. These are integers of the form \code{CCYY}, e.g. 
        \code{2000}.
        }
  
}


\value{

    The function \code{listHolidays} returns a character vector
    with the names of the supported holidays.
    
    The holiday functions return an ISO-8601 formatted 'timeDate'
    of the requested holiday.
   
}


\examples{
## listHolidays -
   listHolidays()
   
## CHSechselaeuten -
   # Sechselaeuten a half Day Bank Holiday in Switzerland
   CHSechselaeuten(2000:2010)
   CHSechselaeuten(currentYear)
   
## German Unification Day:
   DEGermanUnity(currentYear)
}


\keyword{chron}

