\name{length}


\alias{length}
\alias{length.timeDate}


\title{Length of a 'timeDate' Object}


\description{
    
    Returns the length of a 'timeDate' object. 
       
}


\usage{
\method{length}{timeDate}(x)
}


\arguments{
  
    
    \item{x}{
        an object of class \code{timeDate}.
        }
                
}


\value{
   
    returns an integer of length 1.

}


\examples{  
## timCalendar -
   tC = timeCalendar()
   
## length - 
   length(tC)
}


\keyword{chron}

