\name{listHolidays}


\alias{listHolidays}


\title{List of Holidayss}


\description{
    
    Returns the list of holidays.     
                
}


\usage{
listHolidays() 
}


\value{

    returns a list of holidays as a data.frame object.

}


\examples{
## listHolidays -
   listHolidays()
}

\keyword{chron}

