\name{plot-methods}


\title{Plot Methods}


\alias{plot.timeDate}
\alias{points.timeDate}
\alias{lines.timeDate}


\description{

    Plot methods 'timeDate' object.
}


\usage{
\method{plot}{timeDate}(x, y, \dots)
}


\arguments{

    \item{object}{
        an object of class \code{timeDate}.
        }
    \item{\dots}{
        arguments passed to other methods.
        }

}


\value{

    returns a summary report of the details of a \code{timeDate}
    object. This includes the starting and end date, the number of
    dates the format and the financial center in use.
   
}


\examples{
## plot -
}


\keyword{chron}

