\name{show-methods}


\docType{methods}


\alias{show-methods}
\alias{show,ANY-method}
\alias{show,timeDate-method}
\alias{show.timeDate}


\title{Show Methods}


\description{

    Show methods for 'timeDate' objects. 

}


\section{Methods}{
    \describe{

    \item{object = "ANY"}{
        Generic function.
        }
    \item{object = "timeDate"}{
        Print function for objects of class \code{"timeDate"}.
        } 
        
    }
}


\examples{
## print | show -
   print(timeCalendar())
}


\keyword{chron}

