\name{myFinCenter}


\title{myFinCenter Variable}


\alias{myFinCenter}


\description{

    A character string with the name of my 
    financial center.

}


\usage{
myFinCenter
}


\value{

    returns the name of myFinCenter.

}


\examples{
## myFinCenter - the global setting currently used:
   myFinCenter
}


\keyword{chron}

