\name{myUnits}


\title{Frequency of date/time Units}


\alias{currentYear}


\description{

    A variable with the frequency of date/units.

}


\usage{
myUnits
}


\value{

    returns the the date/time units, a acharacter value.
    By default \code{"days"}.

}



\examples{
## myUnits -
   myUnits
}


\keyword{chron}

