\name{RmetricsUtilities}

\alias{RmetricsUtilities}

\alias{modify}
\alias{modify.default}

\alias{atoms}
\alias{atoms.default}

\alias{currentYear}
\alias{myUnits}


\title{Some Rmetrics Utilities}


\description{
    
    A collection and description of functions which
    add to Rmetrics some additional generic functions 
    which may be useful.
    \cr
    
}


\usage{
currentYear
myUnits
}


\details{

    For details we refer to the original help pages.
    
}
    


\author{

    Diethelm Wuertz for the Rmetrics \R-port.
    
}


\examples{
## SOURCE("fCalendar.1A-RmetricsUtilities")
}


\keyword{programming}

